feat(android): implement deep link support for Android
Fixes #19562

Adds the needed manifest info to tell Android which links we supports.
Then adds the Java code necessary to handle it. In this case, we listen to the Intent and if there is a URL, we emit an event that can be listened to in C++, which gets sent to the url_manager in Nim. Then we reuse the same code we already had.