fix(community): disable anyone can be admin or member permissions
Fixes #19620

We enabled users to create "anyone can be admin" or "anyone can join" permissions , which was wrong. The former would be chaotic and makes no sense to support. The latter is just the default of any community without permissions, so it was just cluttering the description.

The solution was to disable the "Community" part of the "is allowed to" popup when "anyone" is active.

I also improved the error and loading handling of the permission creation flow. We can make it async now if we want and we show a better error to the user if something goes wrong. Before, if there was an error, you lost the screen and had to start everything again.